+++ src/App.js
@@ -1,28 +1,28 @@
 import { BrowserRouter, Route, Routes  } from "react-router-dom";
 import styled from "styled-components";
-import Home from "./component/page/Home";
-import User from "./component/page/User";
-import Post from "./component/page/Post";
-import PostDetail from "./component/page/PostDetail";
-import Council from "./component/page/Council";
-import CouncilDetail from "./component/page/CouncilDetail";
-import SignIn from "./component/page/SignIn";
-import SignUp from "./component/page/SignUp";
-import Chat from "./component/page/Chat";
-import InChat from "./component/page/InChat";
-import Admin from "./component/page/Admin";
-import Write from "./component/page/Write";
-import CiManage from "./component/page/CiManage";
-import CiCreate from "./component/page/CiCreate";
-import CiAdd from "./component/page/CiAdd";
-import Certifi from "./component/page/Certifi";
-import CertifiManage from "./component/page/CertifiManage"
-import CouncilManage from "./component/page/CouncilManage"
-import ManagerSignIn from "./component/page/ManagerSignIn";
-import CouncilItemCreate from "./component/page/CouncilItemCreate";
-import CouncilInfoUpdate from "./component/page/CouncilInfoUpdate";
-import Notice from "./component/page/Notice";
-import NoticeDetail from "./component/page/NoticeDetail";
+import Home from "./pages/UserPages/Home";
+import User from "./pages/UserPages/User";
+import PostList from "./pages/UserPages/PostList";
+import PostDetail from "./pages/UserPages/PostDetail";
+import CouncilList from "./pages/UserPages/CouncilList";
+import CouncilDetail from "./pages/CouncilManagerPages/CouncilDetail";
+import SignIn from "./pages/UserPages/SignIn";
+import SignUp from "./pages/UserPages/SignUp";
+import ChatRoomList from "./pages/UserPages/ChatRoomList";
+import Chat from "./pages/UserPages/Chat";
+import AdminManagement from "./pages/AdminPages/AdminManagement";
+import PostEdit from "./pages/UserPages/PostEdit";
+import AdminCouncilManagement from "./pages/AdminPages/AdminCouncilManagement";
+import AdminCouncilCreation from "./pages/AdminPages/AdminCouncilCreation";
+import AdminCouncilItemCreation from "./pages/AdminPages/AdminCouncilItemCreation";
+import Certification from "./pages/UserPages/Certification";
+import AdminCertificationManagement from "./pages/AdminPages/AdminCertificationManagement"
+import CouncilManagement from "./pages/CouncilManagerPages/CouncilManagement"
+import CouncilSignIn from "./pages/CouncilManagerPages/CouncilSignIn";
+import CreationCouncilItem from "./pages/CouncilManagerPages/CreationCouncilItem";
+import EditCouncilInfo from "./pages/CouncilManagerPages/EditCouncilInfo";
+import NoticeList from "./pages/UserPages/NoticeList";
+import NoticeDetail from "./pages/UserPages/NoticeDetail";
 
 const MainLayout = styled.div`
 `;
@@ -33,29 +33,34 @@ function App() {
     <MainLayout>
       <BrowserRouter>
         <Routes>
+          {/* 유저 */}
           <Route path="/" Component={Home} />
-          <Route path="/certification" Component={Certifi} />
-          <Route path="/council/signin" Component={ManagerSignIn} />
-          <Route path="/council/manage" Component={CouncilManage} />
-          <Route path="/council/manage/info" Component={CouncilInfoUpdate} />
-          <Route path="/council/manage/item" Component={CouncilItemCreate} />
-          <Route path="/admin" Component={Admin} />
-          <Route path="/admin/cimanage" Component={CiManage} />
-          <Route path="/admin/cimanage/create" Component={CiCreate} />
-          <Route path="/admin/certimanage" Component={CertifiManage} />
-          <Route path="/admin/cimanage/add/:id" Component={CiAdd} />
+          <Route path="/certification" Component={Certification} />
           <Route path="/user/:userid" Component={User} />
-          <Route path="/write" Component={Write} />
-          <Route path="/post/:location" Component={Post} />
+          <Route path="/write" Component={PostEdit} />
+          <Route path="/post/:location" Component={PostList} />
           <Route path="/post/:location/:id" component={PostDetail} />
-          <Route path="/council/:campus" Component={Council} />
+          <Route path="/council/:campus" Component={CouncilList} />
           <Route path="/council/:campus/:id" Component={CouncilDetail} />
           <Route path="/signin" Component={SignIn} />
           <Route path="/signup" Component={SignUp} />
-          <Route path="/chat" Component={Chat} />
-          <Route path="/chat/:metype/:id/:other" Component={InChat} />
-          <Route path="/notice" Component={Notice} />
+          <Route path="/chat" Component={ChatRoomList} />
+          <Route path="/chat/:metype/:id/:other" Component={Chat} />
+          <Route path="/notice" Component={NoticeList} />
           <Route path="/notice/:id" Component={NoticeDetail} />
+          {/* 어드민 */}
+          <Route path="/admin" Component={AdminManagement} />
+          <Route path="/admin/cimanage" Component={AdminCouncilManagement} />
+          <Route path="/admin/cimanage/create" Component={AdminCouncilCreation} />
+          <Route path="/admin/certimanage" Component={AdminCertificationManagement} />
+          <Route path="/admin/cimanage/add/:id" Component={AdminCouncilItemCreation} />
+          {/* 학생회 */}
+          <Route path="/council/signin" Component={CouncilSignIn} />
+          <Route path="/council/manage" Component={CouncilManagement} />
+          <Route path="/council/manage/info" Component={EditCouncilInfo} />
+          <Route path="/council/manage/item" Component={CreationCouncilItem} />
+          
+        
         </Routes>
       </BrowserRouter>
     </MainLayout>
