+++ src/components/layout/MessageInputBar.jsx
