+++ src/pages/AdminPages/AdminCouncilManagement.jsx
@@ -4,58 +4,17 @@
 사용법: App.js에서 라우팅됨.
 기타: ADMIN 권한 유저만 접근 가능
 */
-import { Link, useNavigate,useParams } from "react-router-dom";
-
+import { Link, useNavigate } from "react-router-dom";
 import { useState, useEffect } from "react";
 import { useCookies } from "react-cookie";
 import axios from "axios";
 import styled from "styled-components";
-import Header from "../layout/Header";
+import Header from "../../components/layout/Header";
 
 
-const CouncilBox = styled.div`
-    position: absolute;
-    width: 100%;
-    /* padding: 20px; */
-    height: 100%;
-    max-width: 700px;
-    background: #ffffff;
+const AdminBox = styled.div`
+    padding: 0 20px;
 `;
-
-const SubTitle = styled.div`
-   font-size: 18px;
-    font-weight: 800;
-    color: #505050;
-    margin-bottom: 10px;  
-    
-`;
-
-const CampusAnnoBox = styled.div`
-   background: #ffffff;
-    margin-bottom: 20px;
-    border-radius: 15px;
-    box-shadow: rgba(215, 218, 220, 0.5) 0px 0px 15px;
-    padding: 20px 10px;
-    font-size: 18px;
-    font-weight: 800;
-    color : #BCC4D1;
-    & div{
-      font-size: 28px;
-      color : #9e9e9e;
-      margin-bottom: 5px;
-    }
-    & div span{ 
-      font-size: 28px;
-      color : #379DFF;
-    }
-`;
-
-const CouncilContents = styled.div`
-
-    padding: 20px;
-
-`;
-
 
 const CollegeBox = styled.ul`
     background: #ffffff;
@@ -96,7 +55,7 @@ const RemoveBtn = styled.button`
 `;
 
 
-const CouncilBlock = styled.li`
+const Council = styled.li`
     list-style: none;
     padding: 15px 10px; 
     line-height: 25px;
@@ -138,14 +97,11 @@ const CreateCouncil = styled.div`
 `;
 
 
-const Council = () => {
+const AdminCouncilManagement = () => {
     const [groupedCouncilList, setGroupedCouncilList] = useState([]); // 채팅방 리스트 상태
     const [key, setKey] = useState(0); 
     const [cookies] = useCookies(); // 쿠키 사용하기 위해
     const navigate = useNavigate(); // 페이지 이동 위해
-    const { campus } = useParams(); 
-    console.log(campus);
-
   
     useEffect(() => {
       const fetchCouncils = async () => {
@@ -200,37 +156,38 @@ const Council = () => {
     };
 
     return (
-        <CouncilBox>
-            <Header headerType={"detail"} headerText={"학생회"}></Header>
-            
-            <CouncilContents>
-              <SubTitle>총 35개의 학생회에서 물품대여중🫶</SubTitle>
-              <CampusAnnoBox>
-                <div><span>{campus == 'g' ? "글로벌" : "메디컬"}</span> 캠퍼스입니다🙂</div>
-                <span>실시간</span>으로 물건 개수를 확인하세요!
-              </CampusAnnoBox>
-              {groupedCouncilList.map((college, index) => (
-                  <div key={index}>
-                      <CollegeName>{college[0] != null ? college[0].college.slice(1) : null}</CollegeName>
-                      <CollegeBox>
-                          {college.map((council) => (
-                             <Link to={"/council/"+campus+"/"+council.councilId} >
-                              <CouncilBlock key={council.councilId}>
-                                  <CouncilImg></CouncilImg>
-                                  <CouncilContent>
-                                      <CouncilName>{council.name}</CouncilName>
-                                      <ItemInfo>제공 물품 {council.providedItemCount} 대여 물품 {council.rentalItemCount}</ItemInfo>
-                                  </CouncilContent>
-                              </CouncilBlock>
-                             </Link>
-
-                          ))}
-                      </CollegeBox>
-                  </div>
-              ))}
-            </CouncilContents>
-        </CouncilBox>
+        <AdminBox>
+            <Header headerType={"admin"} headerText={"학생회 대여품 관리"}></Header>
+            <Link to={"/admin/cimanage/create"}>
+                <CreateCouncil>
+                    <img src={"/image/write_black.svg"}></img>
+                </CreateCouncil>
+            </Link>
+            <div>
+            {groupedCouncilList.map((college, index) => (
+                <div key={index}>
+                    <CollegeName>{college[0] != null ? college[0].college.slice(1) : null}</CollegeName>
+                    <CollegeBox>
+                        {college.map((council) => (
+                            <Council key={council.councilId}>
+                                <CouncilImg></CouncilImg>
+                                <Link to={"/admin/cimanage/add/"+council.councilId} >
+                                <CouncilContent>
+                                    <CouncilName>{council.name}</CouncilName>
+                                    <ItemInfo>제공 물품 {council.providedItemCount} 대여 물품 {council.rentalItemCount}</ItemInfo>
+                                </CouncilContent>
+                                </Link>
+                                <RemoveBtn onClick={() => removeCouncil(council.councilId)}>
+                                    <img src={"/image/remove.svg"}></img>
+                                </RemoveBtn>
+                            </Council>
+                        ))}
+                    </CollegeBox>
+                </div>
+            ))}
+            </div>
+        </AdminBox>
     );
   };
 
-export default Council;
+export default AdminCouncilManagement;
