+++ src/pages/AdminPages/Register.jsx
@@ -8,35 +8,16 @@ import React, { useRef, useState } from "react";
 import { Link, useNavigate } from "react-router-dom";
 import styled from "styled-components";
 import axios from "axios";
-import Footer from "../layout/Footer";
+import Footer from "../../components/layout/Footer";
 
 // 회원가입 form 박스
 const SignUpBox = styled.div`
     margin: 20px auto 0 20px;
-    padding: 70px 0px;
+    /* padding: 70px 0px; */
     width: 90%;
     /* background : #ffffff; */
     border-radius: 20px;
     text-align: center;
-`;
-
-// 타이틀
-const Title = styled.div`
-    text-align: center;
-    height: 45px;
-    line-height: 45px;
-    margin-bottom: 120px;
-    font-size: 60px;
-    font-weight: 850;
-    color : #379DFF;
-`;
-
-// 서브 타이틀
-const SubTitle = styled.div`
-    text-align: center;
-    font-size: 18px;
-    font-weight: 600;
-    color : #adb5c2;
 `;
 
 // 아이디 패스워드 조건 사항
@@ -122,7 +103,7 @@ const FileInputBox = styled.input`
     }
 `;
 
-const SignUp = () => {
+const Register = () => {
   const navigate = useNavigate(); // 페이지 이동
 
   // 입력박스 자동 포커스를 위해
@@ -261,87 +242,14 @@ const SignUp = () => {
   return (
     <div>
       <SignUpBox>
-        <Title>
-          <SubTitle>
-            당장 필요할때 바로 빌리자
-          </SubTitle>
-          대학빌림
-        </Title>
-
-        {/* 별명 */}
-        <InputBox
-          type="text"
-          ref={nicknameRef}
-          name="nickname"
-          value={nickname}
-          placeholder="별명"
-          onChange={(e) => {
-            setNickname(e.target.value);
-          }}
-          onKeyDown={(e) => { activeEnter(e) }} />
-        <Requirements>최대 10글자</Requirements>
-
-        {/* 이메일(제작중) */}
-        {/* <InputBox
-          type="email"
-          // ref={nicknameRef}
-          name="email"
-          // value={nickname}
-          placeholder="이메일"
-          // onChange={(e) => {
-          //   setNickname(e.target.value);
-          // }}
-          onKeyDown={(e) => {activeEnter(e)}}/> */}
-
-        {/* 아이디 */}
-        <InputBox
-          type="text"
-          ref={useridRef}
-          name="id"
-          value={userid}
-          placeholder="아이디"
-          onChange={(e) => {
-            setUserid(e.target.value);
-          }}
-          onKeyDown={(e) => { activeEnter(e) }}
-        />
-        <Requirements>영문 소문자, 숫자 조합 5~15자리</Requirements>
-
-        {/* 비밀번호 */}
-        <InputBox
-          type="password"
-          ref={passwordRef}
-          name="password"
-          value={password}
-          placeholder="비밀번호"
-          onChange={(e) => {
-            setPassword(e.target.value);
-          }}
-          onKeyDown={(e) => { activeEnter(e) }} />
-
-        {/* 비밀번호 확인*/}
-        <InputBox
-          type="password"
-          ref={password2Ref}
-          name="password2"
-          value={password2}
-          placeholder="비밀번호 확인"
-          onChange={(e) => {
-            setPassword2(e.target.value);
-          }}
-          onKeyDown={(e) => { activeEnter(e) }} />
-        <Requirements>대소문자, 숫자, @!? 조합 5~15자리</Requirements>
 
         {/* 이미지 형식 제한해야함 */}
         <FileInputBtn for="file">
-          <div>프로필 사진 추가</div>
+          <div>학생회 사진 추가</div>
         </FileInputBtn>
         <FileInputBox type="file" name="file" id="file" onChange={handleFileChange} />
         {/* 제출 버튼 */}
-        <SubmitBtn onClick={handleSignUp}>회원가입</SubmitBtn>
-        <Link to={"/signin"}>
-          <GoToSignIn>로그인 하러가기</GoToSignIn>
-        </Link>
+        {/* <SubmitBtn onClick={handleSignUp}>회원가입</SubmitBtn> */}
 
       </SignUpBox>
       <Footer></Footer>
@@ -349,4 +257,4 @@ const SignUp = () => {
   );
 };
 
-export default SignUp;
+export default Register;
