+++ src/pages/CouncilManagerPages/CouncilManagement.jsx
@@ -9,7 +9,7 @@ import { useState, useEffect } from "react";
 import { useCookies } from "react-cookie";
 import axios from "axios";
 import styled from "styled-components";
-import Header from "../layout/Header";
+import Header from "../../components/layout/Header";
 
 
 const AdminBox = styled.div`
@@ -129,7 +129,7 @@ const Tag = styled.span`
   margin-right: 10px;
 `;
 
-const CouncilManage = () => {
+const CouncilManagement = () => {
     const [councilData, setCouncilData] = useState({items:[]}); // 채팅방 리스트 상태
     const [key, setKey] = useState(0); 
     const [cookies] = useCookies(); // 쿠키 사용하기 위해
@@ -245,4 +245,4 @@ const CouncilManage = () => {
     );
   };
 
-export default CouncilManage;
+export default CouncilManagement;
